shared subscription stream info object
Summary:
For bounding subscription serve queues by amount of memory used by queued object,
we need to track the metric at both ends of the queue.
Introducing a SubscriptionStreamInfo object that will be accessible from both ends of
the queue.

In next diffs we will update subscribe_*() APIs to allocate and use this shared object.

Differential Revision: D87018141

fbshipit-source-id: f697af5cba1a2d96671a8ddc2ee6fa6641e568e4