Set SaiPhyRetimer HwInfo with XPHY ID for PAI SDK
Summary:
As titled, configure HwInfo attribute for Broadcom PAI SDK retimer in the create attributes

Previously, HwInfo was set to std::nullopt, After looking at the example code and internal PAI code, the SDK actually uses  hwinfo to match against corresponding phy id in the config file, so we need to set this.

More context on this:

1. **Switch Creation & Storage**: During switch creation, the SDK extracts the HwInfo value and stores it internally in its PHY instance structure (phy_info.phy_addr).

2. **Config File Matching**: The SDK uses the stored HwInfo value to locate the corresponding configuration row in the BCM config file. It matches HwInfo against the phy_id field in the config to retrieve corresponding rows.

This change encodes the  XPHY ID (basically xphyid ranging from 0-31 in our case) as a 4-byte little-endian integer and passes it via the HwInfo attribute. The encoding matches Broadcom's reference implementation (checked the example implementation code)

Reviewed By: harshitgulati18

Differential Revision: D87834464

fbshipit-source-id: c02e4c60fbb172b0010cdcb5287c35bc5dae2baa