Set SaiPhyRetimer profile ID to XPHY ID for future flexibility
Summary:
As titled.

Currently, this update is not strictly necessary. The SDK provides us with a profile ID to retrieve information such as warmboot decisions and configuration files, but at present, we return the same information regardless of the XPHY.

But I still think this is a good idea to assign unique profile IDs now for two main reasons:
1) This enables support for future use cases, such as per-retimer configuration files, and allows us to track which profile ID (aka which xphy) the SDK is querying. Makes easier to debug.
2) The example code of PAI uses different profile IDs, so aligning with this approach is a safer choice.

Reviewed By: harshitgulati18

Differential Revision: D87835927

fbshipit-source-id: 645aafe3fa3531c84c2ee230437a0d00e383a624