Configure Broadcom PAI SDK config file path in SaiPhyPlatform
Summary:
Add config file path  for Broadcom PAI SDK initialization.

The SDK will ask for this information through the callback function pointer we provide.

We verbally communicated that this change would be needed to Celestica, but better to put the code here, and they can change the path as per their setup (I will give a ping to them after this is landed, in case they already have not implemented something like this locally)

For the time being, just setting the file path to be "/lib/firmware/fboss/phy_config" : We need to revisit this afterwards.

Reviewed By: harshitgulati18

Differential Revision: D87838987

fbshipit-source-id: 18db63fe42098c93764ac6f2f63a90db7f59678e