Remove device_addr checks in read/write callback functions
Summary:
We previously assumed that the SDK would  pass the device_addr as PmdDeviceAddr (1) in these functions (similar to credo), but after looking at the SDK’s internal code, that turned out not to be the case.

What actually happens is that the SDK passes the value of HwInfo as the device_addr parameter rather. I’m removing the check which most likely would crash for anacapa.

As for why the SDK passes this  value, I am not entirely sure. In our setup, we don’t need a separate identifier for the XPHY in these functions, since the passed platform_context already points  to the retimer object for that XPHY and we can directly use that. Maybe in some other design or use case, having this extra information would be useful, and the SDK is just providing it for flexibility. That’s just my guess though, I can check with the Broadcom team later on this.

Also, since we’re not using the device_addr value, removing the check is not complicating anything.

Reviewed By: harshitgulati18

Differential Revision: D87839993

fbshipit-source-id: 685121b6f01963f1a6f16115987c4501854f655f