Only process latest DsfSubscription updates in the queue
Summary:
This is to fix the race condition where GR expiry processing could lead to clearing out of neighbor entries.

**Problem**
Imagine state update queue is busy and following events happened
1. Remote device learned new neighbor (update 1)
2. GR expiry happened
3. Remote device send new updates to the device after GR (update 2)

Since we have the optimization [nextDsfUpdate](https://fburl.com/code/letw7y1j), nextDsfUpdate will store update 1. Hence the queue would look like
1. [Update1 Event]
2. [Update1 Event, GR expiry]
3. [Update1 Event, GR expiry] <-- nextDsfUpdate stores update 1 previously, hence [we don't scheduled a new update](https://fburl.com/code/b3au21ex) but rather update nextDsfUpdate to store update2.

The net result is that we process update 2 in the update 1 event in the queue, but we then clearing it out after processing GR expiry.

**Attempted fix #1 D85073824**
[Insufficient] When queueing GR expiry, also clear out nextDsfUpdate such that we can enqueue new event from update2.

Now the queue and nextDsfUpdate looks like this
 1. Queue = [Update1 Event], nextDsfUpdate = update1
2. Queue = [Update1 Event, GR expiry], nextDsfUpdate = nullptr <-- because of the fix
3. Queue = [Update1 Event, GR expiry, Update2 Event], nextDsfUpdate = update2

This is not sufficient because, when we process Update1 Event, it will actually use `nextDsfUpdate = update2` and then set nextDsfUpdate to nullptr. Therefore,
- Dequeue Update1 Event, Queue = [GR expiry, Update2 Event]. It will process `nextDsfUpdate` which is update2, then reset `nextDsfUpdate` to nullptr.
- Dequeue GR expiry, Queue = [Update2 Event]. This will wipe out all the entries.
- Dequeue Update2 Event, nextDsfUpdate is already nullptr and hence no update.

Which gets us into the same problem state.

**This fix**
Only the last update event in the queue should actually process the event. In this specific example of `Queue = [Update1 Event, GR expiry, Update2 Event], nextDsfUpdate = update2`, update 1 event should not process the update2 event until it is update2's turn.

To achieve this, keep track of a sequence number of the update. We only process the update if the current sequence number is equal to the latest sequence number.

Reviewed By: nivinl

Differential Revision:
D87839832

Privacy Context Container: L1297311

fbshipit-source-id: ecd52fffcafd784e10b0922ed07eb2d94499fd4c