Add changelog plugin
Summary:
Purely vibecoded feature - I am walking into the future with  an open mind.

Motivation: This was one of the main asks that I heard from Accton during  the wedge800 bringup. Without this it is pretty hard to keep up  with the changes to documentation. They are generally pretty active at reading things and staying on top of it, but without notifications like this it is easy to miss things.

Feature: A simple docusaurus plugin that reads git log to get all of the commits that affected files in the docs/ directory and creates a changelog page in the documentation displaying all of those commits in chronological order.
* Commits are grouped by month
* Includes link to the actual commit

Note: This will only build successfully when run from a git checkout. If building internally on a devserver it will fail and the changelog page will be empty. This is fine since the actual page will be built on github. The rest of the documentation still builds successfully.

Reviewed By: KevinYakar

Differential Revision: D87877259

fbshipit-source-id: 769dc6ba1c06656abd12a6bfa3110eb1884a647c