Preventing high power mode setting once module is already in high power mode
Summary:
The diff handles the special handling for the finisar 200G_FR4 optics. The special handling is required for finisar optics as we have observed on the fleet on minipack2 with finisar operating in 100G,  that Tx-LOL alarm is consistent on the host side. We are coming with the below seqeuence based on the multiple iteration of various programming sequence. Finisar is actively debugging the issue for potentail f.w fix.
-> The module is kept at low power mode in the ready state machine state
-> Utility functions has been introduced for better code structuring
-> 200G_FR4 optics is moved to high power mode only after completion of AppSel programming
-> Poll is introduced for verifying the module in ready state.
-> Added the logic for checking the module state before resetting the lower power bit in customizetransceiverlocked function.

Reviewed By: harshitgulati18

Differential Revision: D87838300

fbshipit-source-id: f27c6d313747f02eec5a36ce61a6fa288c0675d6