Back out "Update UT to also check map size"
Summary:
Original commit changeset: 5786f4ab1fed

Original Phabricator Diff: D87851442

D87839832 introduced a race that caused S599716.

There is race condition introduced in the intended fix. The increment of lastUpdateSeqNum_ should only happen if the update requires scheduling. Reverting the diff to go back to previous state.

Reverting the UT first as it would fail because of the revert.

Reviewed By: ashutoshgrewal, shri-khare

Differential Revision: D88545837

fbshipit-source-id: 0502345faf1c1e92bb97df12a7e7a822ef0cce3a