Generate I2C replayer log with i2ctransfer command
Summary:
Today we are generating a text file with i2c replayer, which is useful to visualize in a CSV.

A replayer log with i2cget/i2cset/i2ctransfer commands is much more easier to replay for us and the vendors. Thus generating such a replayer in this diff.

We need to know the i2c bus number and that's the only thing user would need to add (replace #i2cBus with actual i2c bus number) before running the replayer

Reviewed By: birdsoup

Differential Revision:
D88030240

Privacy Context Container: L1390529

fbshipit-source-id: de6841f2fdea2ba4dab17a0d4a32b58cf891dd2a