Match requirements to test case names
Summary:
Our next goal for documents is to be able to link between test cases and the item in the specification that each test tests.

I'm testing out something here that I think could be cool, but I'm not sure if it's worth getting away from pure markdown.

Docusaurus supports MDX which is just inline react components rendered in document. This diff creates a little "button" in front of a test case item, and when you hover over it it shows the name of the test case.

It's pretty easy to put into the source:

```
<TestCase id="CDEV.SupportsIOCTL">
Text content of the requirement
  * including any other formatting
</TestCase>
```

And then you end up with something like this:
{F1981026245}

Reviewed By: ezeob002

Differential Revision: D79845231

fbshipit-source-id: 828aad6333c4fd5f11e20faae7dddad8634340c8