Fix fan_led regex
Summary: Celestica sent a PR: D88685458 but I didn't like the way they fixed this issue, so I did it here. The regex was failing on "fan_led" because it tried to convert an empty string into an int which failed. Simply capture into a string instead of an int and nothing else needs to change for the test.

Reviewed By: ezeob002

Differential Revision: D88689340

fbshipit-source-id: 7dbbd8feaa54c0bd7637aac1d3606940d4db02e2