Remove i2cdetect and i2cdump from showtech, as they can potentially make devices go bad on some platforms
Summary:
[FBOSS/Platform] Remove i2cdetect and i2cdump from showtech, as they can potentially make devices go bad on some platforms
In Kodiak3, SiLab clock generator may be affected by these i2c detect and i2c dump routine, and go to bad state, which takes ASIC down.
This diff removes these lines, and leave remarks in the code, so as to give the context to whoever that may attemp to add these lines back again.

Reviewed By: Scott8440

Differential Revision: D88797151

fbshipit-source-id: ce84c18982b3f39eb08de8ee845fe3e5bd7be48e