Fix agent_ensemble_link test results grouping in known_bad_test_report.py
Summary:
Per discussion with Prasoon, we are switching the source of these results from TestInfraDB table to Fboss Testing table – this is because TestInfraDB table seems to have some data retention policies that don't line up with what we expect (results after 6 days are not stored, and the amount of data stored is variable).

In tandem with D88277581, this fix should enable the proper reporting of known bad tests for agent ensemble link tests.

Differential Revision: D86803899

fbshipit-source-id: 2c4ea33c4106da0b12ce2e9cb5f60804ca696291