Zero-initialize numOfSlot in PhyManager
Summary:
Seems like some UTs were passing in dev mode but not in opt mode because of an uninitialized variable in `PhyManager.h`.

In actual prod / hw device use cases, this variable is set (e.g. https://fburl.com/code/5d61p8ku), but in `MockPhyManager` this is never set, so was likely stalling on some non-zeroed out value in opt mode. So this is a test-specific regression.

Two changes included here:
- Zero'ing out the value by default as a safety precaution.
- Explicitly zero'ing out the value in `MockPhyManager`

This was not caught on-diff because on-diff runs dev mode, but continuous runs on Conveyor run in opt mode.

Reviewed By: srikrishnagopu

Differential Revision: D89120733

fbshipit-source-id: 02e95628c9d4eda13ec03998619154dd0599d604