adding missing regex for xsw role
Summary:
**Summary**
===============

### [FBOSS CLI] Added Missing Regex for XSW Role

This diff addresses a missing regex pattern for the XSW role in the FBOSS CLI. The modification is made in the `CmdShowLldp.cpp` file, where a new regex pattern `xsw_regex` is added to match port descriptions containing "xsw". If a match is found, the function returns a concatenation of `results[2]` and `results[3]`.

Differential Revision: D89203564

fbshipit-source-id: 145152783b2a777dcbc0c9eaeb48c196c1270243