Move .clang-format to github/ so that both fboss/ and common/ can apply the clang-format
Summary:
When helping Benoit from Nexthop to import PR: https://github.com/facebook/fboss/pull/628,
I hit lots of clang-format issue from our phabricator:
 {F1984269569}

This is due to the .clang-format only exist in github:fboss/fboss directory but not in github:fboss/common
As we ship our fbcode fboss to github, we split the two folder as below
https://www.internalfb.com/code/fbsource//fbcode/opensource/fbcode_builder/manifests/fboss?lines=48-51

Therefore, to make sure we honor the same clang-format, I think we should just move the current fboss/.clang-format to github/ in fbcode,
as github/ will be the root directory in github.
Besides, this .clang-format is mainly used for github formatting, we don't need this for our internal. The first diff to introduce this .clang-format is from D62764920

Reviewed By: kevin645

Differential Revision: D89426180

fbshipit-source-id: be37e6f32e7cb29ab8cb277e9cc8958ab75113ca