enable fboss2 CLI tests in the OSS build (#628)
Summary:
**Pre-submission checklist**
- [x] I've ran the linters locally and fixed lint errors related to the files I modified in this PR. You can install the linters by running `pip install -r requirements-dev.txt && pre-commit install`
- [x] `pre-commit run`

The `fboss2` CLI has a number of unit tests, however none of them are being built/run in the OSS build. Turn the CLI into a library in the cmake build, just like how it’s done in the Buck build, and add a new cmake file for the unit tests that leverages this library.

Various unit tests were including `nettools/common/TestUtils.h`, which is not open source, however they generally didn’t need that header and only needed a header from thrift from the most part.

Some simple stubs were incomplete. The BGP server is taken out when building with `IS_OSS`. This change tries to keep the OSS build as close as possible to the internal build, with minimal usage of `#ifdef`s.

Pull Request resolved: https://github.com/facebook/fboss/pull/628

Test Plan: This PR is only enabling new existing unit tests in the OSS build.

Reviewed By: shiva-menta

Differential Revision: D89410954

Pulled By: joseph5wu

fbshipit-source-id: ee163fdba31b2a0d6768cd927681daba27d4c303