Enable opt-asan mode for FSDB fbpkg
Summary:
By default mode, build mode is opt and so until now, we had the ability to only build fsdb FBpkg in opt mode. This diff introduces the ability to build fsdb fbpkg in opt-asan mode as well

The change adds buck_modes configuration to specify both 'opt' and 'opt-asan' build modes for the fbpkg target.

Differential Revision: D89390920

fbshipit-source-id: 95e0224a7b9308a755245caa7578c4a642233942