[Nexthop] Split FBOSS packaging
Summary
======

Split the single package into five packages based on three targets:

forwarding-stack: The forwarding stack and files necessary for
    execution
platform-stack: The platform stack and supporting files
agent-benchmarks: The agent benchmark HW tests

The platform-stack and forwarding-stack targets have a production
archive and a -test archive which contains the tests and utility files
for that component.

This is necessary so we can install the Platform Stack and Forwarding
Stack at different times in the Distro Image. Further, we don't want
to install the tests in the Distro Image which is supposed to
represent a production-like environment.

This does slightly less than package-fboss.py which copies several
unnecessary files (such as .a library archives) and attempts to
dynamically package library dependencies from the build container. We
don't want the former files installed into the Distro Image and the
latter libraries are currently being put into the Distro Image in
another way.

Test Plan
=========

Inside the build container, run both package.py for all three targets
and manually verify that each produces the matching tar files.