Fix some timing assumptions in PortStateMachineTest
Summary:
We want to remove the assumption that all PortIDs have to be tied to a TransceiverID. PortManager is almost at this point, but there are a few data structure / func signature changes that would need to be made to make this work. This stack will do that.

When I started making changes to these functions, I started seeing some UT failures and realized that there are some unsafe assumptions being made about execution order of events, particularly wrt to phy programming.

This diff refactors our UTs so we don't assume the specific port execution order of events, given they run in an async context.

Reviewed By: harshitgulati18

Differential Revision: D89614351

fbshipit-source-id: 3a87630a877aac2debe83c06aca830bbacd0cece