Optimization of DsfUpdate queue to not process if not the latest update
Summary:
Add optimizations to skip processing dsf updates if it's not at the end of the queue. This means GR expiry had happended after this update being queued, and there is already a new update after GR expiry.

In this case, no need to process the updates but wait for newer updates to install RIFs and system ports.

Differential Revision:
D89500951

Privacy Context Container: L1297311

fbshipit-source-id: 882039aee195bc2a3d5761ec93318f55de896457